+++ plugins/scaffolder-actions/scaffolder-backend-module-utils/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @roadiehq/scaffolder-backend-module-utils
 
+## 1.9.0
+
+### Minor Changes
+
+- 8d0a7db3: Add new scaffolder action roadiehq:utils:fs:replace to add ability to replace text in files.
+
+## 1.8.10
+
+### Patch Changes
+
+- 060e3b22: Minor documentation changes to JSONata actions for spelling and consistency wrt capitalization and punctuation
+- ec38b255: correct jsonata action tests' descriptions
+
 ## 1.8.9
 
 ### Patch Changes